requirement.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
pymysql
pydantic-settings
python-jose
bcrypt==3.2.2
passlib[bcrypt]
"pydantic[email]"